+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.3.5 [2020-07-29]
 
-### Adding Mac support and pester tests
+### Adding Mac support and more pester tests
 
 Mac support added, as well as more pester tests to check adding and removing of terraform installations.